requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Pillow==6.2.0
requests==2.22.0
tinify==1.5.1
urllib3==1.25.3
